pyproject.toml
[tool.poetry]
name = "serenity.sdk.python"
version = "0.0.0"
authors = ["Cloudwall Support <support@cloudwall.tech>"]

description = "Python SDK for the Serenity digital asset risk API"
readme = "README.md"
license = "MIT"
packages = [
    { include = "serenity_sdk/**/*.py", from = "src/python" }
]

[tool.poetry.dependencies]
azure-identity = "^1.10.0"
bidict = "^0.22.0"
fire = "^0.4.0"
pandas = "^1.4.2"
python = ">=3.8,<4"
requests = "^2.18.4"
scipy = "^1.7.3"
seaborn = "^0.11.2"
serenity-types-python = "0.0.27"
plotly = "^5.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
flake8-rst-docstrings = "^0.2.7"
ipykernel = "^6.16.0"
ipywidgets = "7.7.2"
nbstripout = "^0.6.0"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
sphinx = "^5.2.3"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core==1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."src/python/serenity_sdk/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'